Internal components moved to their own folder.
