+++ src/sections/about/components/socialcard/SocialCard.model.ts
