+++ src/sections/about/data/socialIcons.ts
@@ -1,7 +1,7 @@
-import linkedinIcon from "../../assets/linkedin.svg";
-import outlookIcon from "../../assets/outlook.svg";
-import githubIcon from "../../assets/github.svg";
-import SocialIconModel from "./socialicon/SocialIcon.model";
+import linkedinIcon from "../../../assets/linkedin.svg";
+import outlookIcon from "../../../assets/outlook.svg";
+import githubIcon from "../../../assets/github.svg";
+import SocialIconModel from "../components/socialicon/SocialIcon.model";
 
 export const linkedinSocialIcon: SocialIconModel = {
     href: "https://www.linkedin.com/in/maximiliano-ponce-marquez-a6094b216/",
